Add NDJSON stream type.
Add utilities for working with NDJSON streams. Rather than waiting on streamed NDJSON data all to arrive and deserializing it into an array, an NDJsonStream may simply be polled for individual values, so users do not need to wait for all data to be received before using the data received.