Oh, egads! My data structure is ruined!
But what if... I were to rename the "name" field to "username" so that I can use it for Top10s and disguise it as "name" using an alias? Delightfully devilish, Seymour...

Hacky fix to get around https://github.com/serde-rs/serde/issues/1504. Alias and name should be swapped once the issue is fixed. Also I forgot to make a field public for some reason.